refactor v2 model tests
